test(logic): split diplomatic validation tests for size gate
Break the oversized diplomatic OrderEngine test into focused relation/overture and aid/subsidy suites, and extract shared GP/minor setup helpers for reuse. This keeps the file-size reduction effort for #2216 moving while preserving test behavior.
